implement varcontainer
