+++ src/hoil_server/scripts/hoil_exec_node_builder.py
@@ -1,14 +1,12 @@
 from hoil_exec_nodes import *
-from hoil_utils import VariableTable
+from hoil_utils import VariableTable, ExecVarContainer
 from collections import deque
 import typing
 import shlex
 
 class ExecNodeBuilder:
-    def __init__(self, varTable:VariableTable, loopStack, robot):
-        self.table = varTable
-        self.loopStack = loopStack
-        self.robot = robot
+    def __init__(self, container:ExecVarContainer):
+        self.container = container
 
     def Run(self, stack: deque) -> typing.Optional[ExecNode]:
         # Override this
@@ -18,8 +16,8 @@ class ExecNodeBuilder:
         return shlex.split(stack[0], posix= False)
 
 class DeclNodeBuilder(ExecNodeBuilder):
-    def __init__(self, varTable: VariableTable, loopStack, robot):
-        super().__init__(varTable, loopStack, robot)
+    def __init__(self, container:ExecVarContainer):
+        super().__init__(container)
 
     def Run(self, stack:deque) -> typing.Optional[ExecNode]:
         line = self.GetLine(stack)
@@ -28,14 +26,14 @@ class DeclNodeBuilder(ExecNodeBuilder):
         
         stack.popleft()
         if len(line) == 4:
-            return DeclNode(self.table, line[1], line[2], line[3])
+            return DeclNode(self.container, line[1], line[2], line[3])
         else:
-            return DeclNode(self.table, line[1], line[2], None)
+            return DeclNode(self.container, line[1], line[2], None)
 
 
 class BranchNodeBuilder(ExecNodeBuilder):
-    def __init__(self, varTable: VariableTable, loopStack, robot):
-        super().__init__(varTable, loopStack, robot)
+    def __init__(self, container:ExecVarContainer):
+        super().__init__(container)
 
     def Run(self, stack: deque) -> typing.Optional[ExecNode]:
         line = self.GetLine(stack)
@@ -43,20 +41,20 @@ class BranchNodeBuilder(ExecNodeBuilder):
         if line[0] != '$branch_begin':
             return None
         
-        branchNode = BranchNode()
+        branchNode = BranchNode(self.container)
 
         stack.popleft()
         line = self.GetLine(stack)
         
         while line[0] != '$branch_end':
             if line[0] == '$if' or line[0] == '$elif':
-                condNode = ConditionalNode()
+                condNode = ConditionalNode(self.container)
 
-                cond = ExprNode(self.table, line[1])
+                cond = ExprNode(self.container, line[1])
                 stack.popleft()
 
                 # Will always terminate upon '$*_end'
-                stmt = _BuildExecNode(stack, self.table, self.loopStack, self.robot)
+                stmt = _BuildExecNode(stack, self.container)
                 
                 condNode.condNode = cond
                 condNode.execNode = stmt
@@ -70,7 +68,7 @@ class BranchNodeBuilder(ExecNodeBuilder):
                 stack.popleft()
             elif line[0] == '$else_begin':
                 stack.popleft()
-                node = _BuildExecNode(stack, self.table, self.loopStack, self.robot)
+                node = _BuildExecNode(stack, self.container)
 
                 branchNode.elseNode = node
                 stack.popleft()
@@ -81,8 +79,8 @@ class BranchNodeBuilder(ExecNodeBuilder):
         return branchNode
 
 class ScopedNodeBuilder(ExecNodeBuilder):
-    def __init__(self, varTable: VariableTable, loopStack, robot):
-        super().__init__(varTable, loopStack, robot)
+    def __init__(self, container:ExecVarContainer):
+        super().__init__(container)
     
     def Run(self, stack: deque) -> typing.Optional[ExecNode]:
         line = self.GetLine(stack)
@@ -91,15 +89,15 @@ class ScopedNodeBuilder(ExecNodeBuilder):
             return None
         
         stack.popleft()
-        scopedNode = ScopedNode(self.table)
-        scopedNode.node = _BuildExecNode(stack, self.table, self.loopStack, self.robot)
+        scopedNode = ScopedNode(self.container)
+        scopedNode.node = _BuildExecNode(stack, self.container)
         stack.popleft()
 
         return scopedNode
 
 class LoopNodeBuilder(ExecNodeBuilder):
-    def __init__(self, varTable: VariableTable, loopStack, robot):
-        super().__init__(varTable, loopStack, robot)
+    def __init__(self, container:ExecVarContainer):
+        super().__init__(container)
 
     def Run(self, stack: deque) -> typing.Optional[ExecNode]:
         line = self.GetLine(stack)
@@ -107,22 +105,22 @@ class LoopNodeBuilder(ExecNodeBuilder):
         if line[0] != '$while':
             return None
     
-        cond = ExprNode(self.table, 'true')
+        cond = ExprNode(self.container, 'true')
 
         if len(line) == 2:
-            cond = ExprNode(self.table, line[1])
+            cond = ExprNode(self.container, line[1])
     
         stack.popleft()
 
-        body = _BuildExecNode(stack, self.table, self.loopStack, self.robot)
+        body = _BuildExecNode(stack, self.container)
 
         stack.popleft()
 
-        return LoopNode(self.table, self.loopStack, cond, body)
+        return LoopNode(self.container, cond, body)
 
 class JumpNodeBuilder(ExecNodeBuilder):
-    def __init__(self, varTable: VariableTable, loopStack, robot):
-        super().__init__(varTable, loopStack, robot)
+    def __init__(self, container:ExecVarContainer):
+        super().__init__(container)
     
     def Run(self, stack: deque) -> typing.Optional[ExecNode]:
         line = self.GetLine(stack)
@@ -130,30 +128,30 @@ class JumpNodeBuilder(ExecNodeBuilder):
         
         if line[0] == '$break':
             stack.popleft()
-            return BreakNode(self.loopStack)
+            return BreakNode(self.container)
         elif line[0] == '$continue':
             stack.popleft()
-            return ContinueNode()
+            return ContinueNode(self.container)
 
         return None
 
 
 
             
-def _BuildExecNode(stack: deque, varTable: VariableTable, loopStack, robot) -> typing.Optional[ExecNode]:
+def _BuildExecNode(stack: deque, container:ExecVarContainer) -> typing.Optional[ExecNode]:
     """Called recursively to create linked list of execnodes.
     Exit upon encountering end region (if_end, while_end etc)
     as well-formed program will always have a matching pair and therefore a matching func call to handle them"""
     
     builderList = list()
-    builderList.append(DeclNodeBuilder(varTable, loopStack, robot))
-    builderList.append(BranchNodeBuilder(varTable, loopStack, robot))
-    builderList.append(ScopedNodeBuilder(varTable, loopStack, robot))
-    builderList.append(LoopNodeBuilder(varTable, loopStack, robot))
-    builderList.append(JumpNodeBuilder(varTable, loopStack, robot))
+    builderList.append(DeclNodeBuilder(container))
+    builderList.append(BranchNodeBuilder(container))
+    builderList.append(ScopedNodeBuilder(container))
+    builderList.append(LoopNodeBuilder(container))
+    builderList.append(JumpNodeBuilder(container))
 
 
-    head = EmptyNode()
+    head = EmptyNode(container)
     cur = None
 
     while len(stack) > 0:
@@ -184,16 +182,15 @@ def _BuildExecNode(stack: deque, varTable: VariableTable, loopStack, robot) -> t
                 else:
                     cur.next = node
                     cur = node
-
     return head
     
 
 
 
 # Node builder produces a complete list of ExecNodes
-def BuildExecNode(source: str, varTable: VariableTable, loopStack, robot) -> typing.Optional[ExecNode]:
+def BuildExecNode(source: str, container:ExecVarContainer) -> typing.Optional[ExecNode]:
     stack = deque([x for x in source.splitlines() if x != ''])
-    return _BuildExecNode(stack, varTable, loopStack, robot)
+    return _BuildExecNode(stack, container)
